requirements.txt

pandas
prophet
streamlit
apache-airflow
scikit-learn
matplotlib
pyyaml
ortools


